Bug fixing for parsing C
The behaviour with respect to grouping has been changed.
Grouping can be used in two different ways. If the grouping
has non-null `add_function` the previous elements of the
rule where the grouping is used are not passed to the rules
in the grouping as previous children, because it is assumed
that the `add_function` will take care of combining the data
parsed by the rules in the grouping with the previous elements.

Furthermore, this commit contains the following changes:
- Added debugging to `parse_element`.
- Added `tree_name` member to `tree_t` and use instead of
  `type_name` of `tree_node_t` for the name of the tree.
- Added `prev_child_print` function.
- Fixed `print_single_char` escaping of ' or ".
- Added `length` member to `struct string_node_t` needed
  for strings contants that contain '\0' characters.
  The value is equal to size of the allocated memory for the
  string, this including the '\0' termination character.
- Fixed bugs with `string_data_print` and `string_data_add_normal_char`.
- Fixed a reference counting bug in `add_seq_as_list`.
- Many fixed for the C-grammer mostly with respect to
  construction of the Abstract Parse/Syntax Tree.